+++ Sprint1/S1T7_Annotations/src/main/java/n1ex01/OnlineWorker.java
@@ -2,13 +2,13 @@ package n1ex01;
 
 public class OnlineWorker extends Worker{
     private final int INTERNET = 50;
-    public OnlineWorker(String name, String surname, int priceHour) {
+    public OnlineWorker(String name, String surname, double priceHour) {
         super(name, surname, priceHour);
     }
 
     @Override
-    public void calcSalary(int numHours){
-        int pay = numHours * getPriceHour() + INTERNET;
+    public void calcSalary(double numHours){
+        double pay = numHours * getPriceHour() + INTERNET;
         System.out.println("Worker's pay is: " + pay + "eur.");
     }
 }
